hawkeye2ha/hawkeye2ha/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
paho-mqtt>=2.0.0
httpx>=0.27.0
